+++ lib/pages/create/complete_habit_dialog.dart
@@ -12,7 +12,7 @@ import 'package:lotti/get_it.dart';
 import 'package:lotti/logic/persistence_logic.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/widgets/charts/utils.dart';
-import 'package:lotti/widgets/form_builder/cupertino_datepicker.dart';
+import 'package:lotti/widgets/misc/datetime_bottom_sheet.dart';
 import 'package:tinycolor2/tinycolor2.dart';
 import 'package:url_launcher/url_launcher.dart';
 
@@ -37,7 +37,8 @@ class _HabitDialogState extends State<HabitDialog> {
   final JournalDb _db = getIt<JournalDb>();
   final PersistenceLogic persistenceLogic = getIt<PersistenceLogic>();
   final _formKey = GlobalKey<FormBuilderState>();
-  bool dirty = true;
+
+  bool _startReset = false;
 
   final hotkeyCmdS = HotKey(
     KeyCode.keyS,
@@ -50,26 +51,19 @@ class _HabitDialogState extends State<HabitDialog> {
     if (validate()) {
       final formData = _formKey.currentState?.value;
       final habitDefinition = await _db.watchHabitById(widget.habitId).first;
-      final formDate = formData!['date'] as DateTime;
 
       final habitCompletion = HabitCompletionData(
         habitId: widget.habitId,
-        dateTo: formDate == _started && widget.data == null
-            ? DateTime.now()
-            : formDate,
-        dateFrom: formDate,
+        dateTo: !_startReset ? DateTime.now() : _started,
+        dateFrom: _started,
         completionType: completionType,
       );
 
       await persistenceLogic.createHabitCompletionEntry(
         data: habitCompletion,
-        comment: formData['comment'] as String,
+        comment: formData!['comment'] as String,
         habitDefinition: habitDefinition,
       );
-
-      setState(() {
-        dirty = false;
-      });
 
       widget.beamerDelegate.beamBack();
     }
@@ -79,16 +73,15 @@ class _HabitDialogState extends State<HabitDialog> {
   void initState() {
     super.initState();
 
+    DateTime endOfDay() {
+      final date = DateTime.parse(widget.data.toString());
+      return DateTime(date.year, date.month, date.day, 23, 59, 59);
+    }
+
     _started = widget.data != null &&
             widget.data is String &&
             ymd(DateTime.now()) != widget.data
-        ? DateTime.parse(widget.data.toString()).add(
-            const Duration(
-              hours: 23,
-              minutes: 59,
-              seconds: 59,
-            ),
-          )
+        ? endOfDay()
         : DateTime.now();
 
     hotKeyManager.register(
@@ -110,7 +103,7 @@ class _HabitDialogState extends State<HabitDialog> {
     return false;
   }
 
-  late final DateTime _started;
+  late DateTime _started;
 
   @override
   Widget build(BuildContext context) {
@@ -179,11 +172,6 @@ class _HabitDialogState extends State<HabitDialog> {
             child: FormBuilder(
               key: _formKey,
               autovalidateMode: AutovalidateMode.onUserInteraction,
-              onChanged: () {
-                setState(() {
-                  dirty = true;
-                });
-              },
               child: Column(
                 mainAxisSize: MainAxisSize.min,
                 crossAxisAlignment: CrossAxisAlignment.start,
@@ -211,18 +199,32 @@ class _HabitDialogState extends State<HabitDialog> {
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: [
                         inputSpacer,
-                        FormBuilderCupertinoDateTimePicker(
-                          name: 'date',
-                          alwaysUse24HourFormat: true,
-                          format: DateFormat(
-                            "MMMM d, yyyy 'at' HH:mm",
-                          ),
-                          style: newInputStyle().copyWith(color: Colors.black),
+                        TextField(
                           decoration: createDialogInputDecoration(
                             labelText: localizations.addHabitDateLabel,
                           ),
-                          initialValue: _started,
-                          theme: datePickerTheme(),
+                          style: newInputStyle().copyWith(color: Colors.black),
+                          readOnly: true,
+                          controller: TextEditingController(
+                            text: DateFormat(
+                              "MMMM d, yyyy 'at' HH:mm",
+                            ).format(_started),
+                          ),
+                          onTap: () async {
+                            final newDate =
+                                await showModalBottomSheet<DateTime>(
+                              context: context,
+                              builder: (context) {
+                                return DateTimeBottomSheet(_started);
+                              },
+                            );
+                            if (newDate != null) {
+                              setState(() {
+                                _startReset = true;
+                                _started = newDate;
+                              });
+                            }
+                          },
                         ),
                         inputSpacer,
                         FormBuilderTextField(